+++ changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * Add inventory item validation tests.
     * Delete inventory item validation tests.
     * Patch inventory item validation tests.
+    * Edit user validation tests.
 * Models factories
     * Users.
     * Inventory items.